Add closing/opening countdown to store info pane
This commit adds a countdown until closing/opening for stores that
are near closing/opening time. For the moment the countdown until
closing begins 8 hours before the store closes, for testing purposes.
The info pane is still not fully implemented - for the moment it is
just a div with some info in it that has yet to be added to the map.