Fix infowindows to use #infopane node, not html
Infowindows now consume the #infopane node, which means they will change
with the div without needing to redraw the infowindow itself. In practice
this means the displayed time-until-closing will update accurately without
causing a flicker as the infowindow is redrawn. Additionally, the logic
for checking if a store is open was changed to account for stores closed
all day.