Search for SAQs based on lat/lng, not postal code
Instead of looking up stores based on postal code, the PHP script now
searches based on latitude/longitude. This allows results for every
query, fixing the bug where no stores were returned because Google
was returning only the postal code prefix. This change also allows
returning more than just 5 stores in the future, but that is not
implemented in this commit.